// src/app/interfaces/iinvitacion.ts
export interface IInvitacion {
    id: any;
  uuid: any;
  clasificacion: any;
  tipoPases: any;
  numQrsBase: any;
  nombreCompleto: any;
  cargo: any;
  institucion: any;
  email: any;
  numAcompaniantesSeleccionados: any;
  numAcompaniantesTotales: any;
  aceptada: any;
  rechazada: any;
  telefono: any;
  opcionQR: any;
  correoEnviado: any;
}